Move LegacyWidgetPlaceholder to its own file
